Add option -a to install annovar from local file.
This makes it possible to pack the installation routine into a script.
As described in the README.md its possible to install annovar by simply unpacking it, but there is no way to install the GATK version without the interrupt for the annovar url.
Another option would be to add a separate install command for GATK.